backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-core
faiss-cpu
pandas
pypdf
python-docx
openpyxl
funasr
modelscope
torch
torchaudio
